% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_by_position.R
\name{add_by_position}
\alias{add_by_position}
\title{Add element to vector by position}
\usage{
add_by_position(x, y, i)
}
\arguments{
\item{x}{vector}

\item{y}{new element(s) to add to vector. Must be coercible to type of \code{x}}

\item{i}{numeric scalar. Index where to add new elements.
Censored to min = 1 and max = length(x) + 1. Thus to add elements at end of
vector use \code{Inf}}
}
\description{
Add element to vector by position
}
\examples{
add_by_position(LETTERS[1:5], "NEW", 3)

}
