% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Add a new class to object in addition to existing classes}
\usage{
add_class(x, class, position = 1)
}
\arguments{
\item{x}{object to add class to}

\item{class}{string. class to add.}

\item{position}{the position in which to add the class}
}
\value{
object 'x' but with  new class added
}
\description{
Add a new class to object in addition to existing classes
}
\examples{
x <- add_class(tibble(), "NEW", 1)
class(x)

}
