% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_calcs.R
\name{add_zeroes}
\alias{add_zeroes}
\title{Add trailing zeros}
\usage{
add_zeroes(x, len = max(nchar(x), na.rm = TRUE))
}
\arguments{
\item{x}{vector to add trailing zeros}
}
\value{
character vector with traling zeros
}
\description{
Add trailing zeros
}
