% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_to_fun.R
\name{args_to_fun}
\alias{args_to_fun}
\title{Pass arguments from caller environment to another function}
\usage{
args_to_fun(fun)
}
\arguments{
\item{fun}{string referencing function name to pass arguments to.}
}
\description{
This function is useful when you want to pass all named arguments from a function
to another function. I've used this in the pass when you have a wrapper for another
function with the exact same arguments, but may be preferred to ... as it is
more explicit
}
