% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{as_numeric2}
\alias{as_numeric2}
\title{As.numeric but also works for factors}
\usage{
as_numeric2(x)
}
\arguments{
\item{x}{vector to convert to double.}
}
\description{
Wrapper for as.numeric that works the same as as.numeric for everything
except factors for which it returns the levels of the factor (if it can),
rather than the underlying integers
}
\details{
Should work fine as the only (base) methods for as.numeric are
difftime and POSIXlt which don't also have factor as class. Could be problematic
if some class inherits factor, but then provides their own as.numeric method,
which this function would override. could of course get function to check for this
and to avoid, but think it is very unlikely.
}
