% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows2.R
\name{bind_rows2}
\alias{bind_rows2}
\title{Bind rows with same names but different column types}
\usage{
bind_rows2(l)
}
\arguments{
\item{l}{list of dataframes to combind with \code{bind_rows()}}
}
\description{
This is a wrapper around bind_rows() that first coerces columns in different
datasets with the same name but different types. This overrides the default
behaviour in bind_rows in which an error is raised when this occurs.
}
