% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{capitalise_first_letter}
\alias{capitalise_first_letter}
\title{Capitalise first letter of each element of vector}
\usage{
capitalise_first_letter(x, rest_down = TRUE)
}
\arguments{
\item{x}{character vector}
}
\description{
Capitalise first letter of each element of vector
}
