% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{censor_left}
\alias{censor_left}
\alias{censor_right}
\alias{censor_both}
\title{Return censored data}
\usage{
censor_left(x, min)

censor_right(x, max)

censor_both(x, min, max)
}
\arguments{
\item{x}{numeric vector}

\item{min}{numeric scalar describing the minimum value for censoring.}

\item{max}{numeric scalar describing the maximum value for censoring.}
}
\description{
Return censored data
}
\section{Functions}{
\itemize{
\item \code{censor_left()}: left censor data (floor)

\item \code{censor_right()}: right censor data (ceiling)

\item \code{censor_both()}: left and right censor data (floor and ceiling)

}}
