% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date_prose}
\alias{date_prose}
\title{Get dates in prose format}
\usage{
date_prose(x, full_month = TRUE)
}
\arguments{
\item{x}{character vector in yyyy-mm-dd format.}

\item{full_month}{logical whether to return full name of month or
shortened version (e.g,. "Oct")}
}
\value{
characater vector of dates in prose format (e.g., 21 May 1991)
}
\description{
Get dates in prose format
}
