% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_to_categorical.R
\name{dummy_to_categorical}
\alias{dummy_to_categorical}
\title{Convert Dummy Variables to Categorical Variable}
\usage{
dummy_to_categorical(..., levels = NULL, default = NA)
}
\arguments{
\item{...}{A set of logical or binary numeric vectors (or a data frame) representing dummy variables.}

\item{levels}{A character vector specifying the category levels. Defaults to column names.}

\item{default}{The value to assign when all dummy variables in a row are zero. Defaults to \code{NA}.}
}
\value{
A factor variable with levels corresponding to the input categories and the default value.
}
\description{
This function converts a set of dummy (one-hot encoded) variables into a categorical variable.
It assigns a category based on the column with the highest value in each row.
If all values in a row are zero, it assigns a default value.
}
\examples{
dummy_data <- data.frame(A = c(TRUE, FALSE, FALSE),
                         B = c(FALSE, TRUE, FALSE),
                         C = c(FALSE, FALSE, TRUE))
dummy_to_categorical(dummy_data)

dummy_data2 <- data.frame(A = c(1, 0, 0, 0),
                          B = c(0, 1, 0, 0),
                          C = c(0, 0, 1, 0))
dummy_to_categorical(dummy_data2, default = "None")

}
