% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nums_in_strings.R
\name{extract_sci_exp}
\alias{extract_sci_exp}
\title{Extract scientific notation exponent}
\usage{
extract_sci_exp(x)
}
\arguments{
\item{x}{string of numbers in scientific notation}
}
\description{
Extract scientific notation exponent
}
