% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_calcs.R
\name{int2bin}
\alias{int2bin}
\title{Convert integer to binary string}
\usage{
int2bin(x)
}
\arguments{
\item{x}{integer to convert to binary}
}
\value{
character vector
}
\description{
Convert integer to binary string
}
