% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_s3_generics.R
\name{list_s3_generics}
\alias{list_s3_generics}
\title{List all S3 generic function in package}
\usage{
list_s3_generics(
  pkgs = c("base", "utils", "methods", "stats", "graphics", "datasets")
)
}
\arguments{
\item{pkgs}{character vector with names of packages to}
}
\description{
List all S3 generic function in package
}
\examples{
list_s3_generics("base")

}
