% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-calcs.R
\name{msd_label}
\alias{msd_label}
\title{Get M (SD) from numeric vector}
\usage{
msd_label(x, digits = c(2, 2), na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{numeric vector of length 2. Digits to round M and SD to
(in that order). Defaults to \code{c(2, 2)}}

\item{na.rm.}{logical. whether to remove \code{NA}s before calculating M and SD.}
}
\value{
string of form "M (SD)"
}
\description{
Get M (SD) from numeric vector
}
