% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{nice_date}
\alias{nice_date}
\title{Get nice copy}
\usage{
nice_date(x, full_month = TRUE)
}
\arguments{
\item{x}{character vector in yyyy-mm-dd format.}

\item{full_month}{whether to return full name of month or shortened version
(e.g,. "Oct")}
}
\description{
Get nice copy
}
