% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nums_in_strings.R
\name{remove_nums}
\alias{remove_nums}
\alias{extract_nums_tgthr}
\alias{extract_nums_all}
\title{Dealing with strings with numbers}
\usage{
remove_nums(x, trimws = FALSE)

extract_nums_tgthr(x)

extract_nums_all(x)
}
\arguments{
\item{x}{character vector, or other vector which will be coerced to character}

\item{trimws}{logical. whether to call \code{base::trimws()}, trimming all whitespace
at start of end of string before returning.}
}
\description{
Dealing with strings with numbers
}
\section{Functions}{
\itemize{
\item \code{remove_nums}: remove all the numbers from a string

\item \code{extract_nums_tgthr}: extract the numbers from a string all together
(i.e. \verb{12aaa34} -> \code{1234})

\item \code{extract_nums_all}: extract all the separate numbers from a string.
returns list of numeric vectors. (i.e., 12aaa34 -> \verb{[1] 12 [2] 34})
}}

