% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_experiment_funs.R
\name{parse_chatbot_info}
\alias{parse_chatbot_info}
\title{Extract data from Hause's chatbot}
\usage{
parse_chatbot_info(...)
}
\arguments{
\item{...}{Vectors containing the json output from Hause's chatbot}
}
\value{
A nested tibble with the data extracted from the chatbot
}
\description{
This function takes a set of vectors, containing the json output from Hause's
chatbot, parses them and returns a tibble with the data.
}
