% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_calcs.R
\name{pmf}
\alias{pmf}
\title{Create pmf object}
\usage{
pmf(x, p)
}
\arguments{
\item{x}{either a sample of values from which to estimate a PMF based on frequency
if \code{p} is missing, or unique \code{x} values with probability \code{p}}

\item{the}{probabilty of each unqiue value of \code{x} occuring}
}
\value{
\code{data.frame} of class \code{pmf}
}
\description{
Create pmf object
}
\examples{
samp <- sample(c(-1000, 100, 800), size = 1e4, replace = TRUE)
pmf(samp)
pmf(x = c(-1000, 100, 800), p = c(0.25, 0.5, 0.25))
}
