% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_funs.R
\name{quarto_render_to_dir}
\alias{quarto_render_to_dir}
\title{Render Quarto document to another directory}
\usage{
quarto_render_to_dir(
  qmd_file,
  out_dir,
  render_in_temp = FALSE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{qmd_file}{string. file path for input \code{.qmd} file}

\item{out_dir}{string. path to output directory in which to save .qmd output}

\item{render_in_temp}{logical. whether to render in a temporary directory or in
the same directory as the input file. Defaults to \code{FALSE}.}

\item{overwrite}{logical. whether to overwrite existing files in the output
directory.}

\item{...}{passed to quarto_render}
}
\value{
Invisibly returns character vector with output file path(s). Mainly
called for side effects -- rendering quarto in new directory.
}
\description{
Currently \code{quarto::quarto_render()} does not allow to render quarto to another
directory. This function creates a temporary directory within which to render
the .qmd doc, then copies files to the output directory.
}
