% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_funs.R
\name{read_csv_qualtrics}
\alias{read_csv_qualtrics}
\title{Funcion to read csv files downloaded from Qualtrics}
\usage{
read_csv_qualtrics(path, remove_extra_cols = TRUE, ...)
}
\arguments{
\item{path}{character. The path to the file to read.}

\item{remove_extra_cols}{logical. If TRUE, removes columns from qualtrics that
are not often needed for analysis. See source code for list. Default is TRUE.}

\item{...}{additional arguments to pass to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\description{
Wrapper around \code{\link[readr:read_delim]{readr::read_csv()}} to help with reading files from qualtrics.
Automatically moves the second and third rows, keeping column names,
before passing to \code{\link[readr:read_delim]{readr::read_csv()}}.
}
