% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_funs.R
\name{rev_diag}
\alias{rev_diag}
\title{Get Reverse diagonal of a square matrix}
\usage{
rev_diag(n, x = 1)
}
\arguments{
\item{n}{number of rows in matrix. Note, this function forces a square matrix
with n = p.}

\item{x}{value to place in reverse diagonal. defaults to 1.}
}
\description{
Get Reverse diagonal of a square matrix
}
\examples{
rev_diag(3)
rev_diag(5, TRUE)

}
