% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_fun.R
\name{to_fun}
\alias{to_fun}
\title{Pass to function with named arguments, using dynamic dots}
\usage{
to_fun(x, args, fun)
}
\arguments{
\item{x}{first argument of function, normally data}

\item{args}{list of other named arguments to pass to \code{fun}}

\item{fun}{function to pass \code{x} and \code{args} to.}
}
\description{
Pass to function with named arguments, using dynamic dots
}
