% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{today}
\alias{today}
\title{Get today's date in nice format}
\usage{
today(full_month = TRUE)
}
\arguments{
\item{full_month}{whether to return full name of month or shortened version
(e.g,. "Oct")}
}
\description{
Get today's date in nice format
}
\examples{
today()

}
