% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest-vectors.R
\name{unnest_vectors}
\alias{unnest_vectors}
\title{Unnest vectors to data.frame}
\usage{
unnest_vectors(l)
}
\arguments{
\item{l}{list of vectors of same length, although obcsured in levels of nesting}
}
\value{
list of vectors without any nesting.
}
\description{
Unnest vectors to data.frame
}
\examples{
x <- stats::runif(10)
y <- LETTERS[1:10]
l <- list(x = x, y, list(list(x, y), nums = x, y))
unnest_vectors(l)

}
