% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{any0}
\alias{any0}
\alias{all0}
\title{Utility functions}
\usage{
any0(...)

all0(...)
}
\description{
Set of simple utility functions for common use.
}
\section{Functions}{
\itemize{
\item \code{any0}: wrapper for \code{any()} with na.rm = TRUE

\item \code{all0}: wrapper for \code{all()} with na.rm = TRUE
}}

