% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2prose.R
\name{vec2prose}
\alias{vec2prose}
\title{List in prose}
\usage{
vec2prose(x, oxford_comma = TRUE, na.rm = TRUE, warn = TRUE)
}
\arguments{
\item{x}{character vector of length > 0 (after removing \code{NA}s if na.rm = TRUE)}

\item{oxford_comma}{logical. If TRUE (default), output includes oxford comma.}

\item{na.rm}{logical. If TRUE (Default), remove all \code{NA} from output.}

\item{warn}{logical. If TRUE (Default), warn about \code{NA}s in \code{x}.}
}
\value{
character vector of length 1.
}
\description{
Takes a vector and returns prose listing vector elements. Suitable for use in
inline text in Rmarkdown.
}
\examples{
x <- c("dogs", "cats", "pigs")

vec2prose(x)
vec2prose(x, oxford_comma = FALSE)

}
