% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Return environment where function is defined}
\usage{
where(name, env = rlang::caller_env())
}
\arguments{
\item{name}{The name of function to search for. Can be a string, or can be
evaluated with non-standard evaluation.}

\item{env}{environment from which to start recursive search. Defaults to the
caller environment taken from \code{rlang::caller_env()}. Note, setting the \code{env}
argument to anything but the default will nullify the capability for
non-standard evaluation, and string must then be used in \code{name}.}
}
\value{
environment where function is defined.
}
\description{
This is an adaptation of the function written by Hadley Wickham, see
\verb{https://adv-r.hadley.nz/environments.html}. This version allows the user to
specify the function as a string, or to use with non-standard evaluation.
}
\examples{
where(sum)
where("sd")

}
