% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-calcs.R
\name{wilcox_AUC}
\alias{wilcox_AUC}
\title{Get AUC-based effect sizefor wilcoxon rank sum test}
\usage{
wilcox_AUC(x)
}
\arguments{
\item{x}{dataframe output from \code{rstatix::wilcox_test()}}
}
\description{
Get AUC-based effect sizefor wilcoxon rank sum test
}
