% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_funs.R
\name{zeros}
\alias{zeros}
\alias{ones}
\title{Make an array of zeros or ones}
\usage{
zeros(dim)

ones(dim)
}
\arguments{
\item{dim}{dimensions of array of zeros. If scalar returns a vector. If vector input,
returns \code{length(dim)}-dimensional array.}
}
\description{
Make an array of zeros or ones
}
\examples{
zeros(7)
ones(c(3, 4, 2))

}
