% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{censor_left}
\alias{censor_left}
\alias{censor_right}
\title{Return censored data}
\usage{
censor_left(x, min)

censor_right(x, max)
}
\arguments{
\item{x}{numeric vector}

\item{min/max}{numeric scalar describing the min/max for censoring.}
}
\description{
Return censored data
}
\section{Functions}{
\itemize{
\item \code{censor_left}: left censor data (floor)

\item \code{censor_right}: right censor data (ceiling)
}}

