% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{find_dropbox_dir}
\alias{find_dropbox_dir}
\alias{get_dropbox_repo_dir}
\alias{get_dropbox_data_path}
\alias{get_dropbox_output_path}
\alias{get_data_from_dropbox}
\title{Find users Dropbox base directory}
\usage{
find_dropbox_dir(version = c("business", "personal"))

get_dropbox_repo_dir(
  ...,
  repo_name = get_git_repo_name(),
  repos_subdir = "projects",
  dropbox_base_dir = find_dropbox_dir(),
  create_dir = FALSE
)

get_dropbox_data_path(
  ...,
  repos_subdir = "projects",
  repo_name = get_git_repo_name(),
  dropbox_base_dir = find_dropbox_dir(),
  create_dir = FALSE
)

get_dropbox_output_path(
  ...,
  repos_subdir = "projects",
  repo_name = get_git_repo_name(),
  dropbox_base_dir = find_dropbox_dir(),
  create_dir = FALSE
)

get_data_from_dropbox(file_name, ...)
}
\arguments{
\item{version}{A string specifying the version of Dropbox to use, if 2 are
found on he machine. Can be either "personal" or "business" (default).}

\item{...}{character vector. Additional path components to be joined to the
end of the path (before 'data', or 'output' are appended for \code{get_dropbox_data_path()}
and \code{get_dropbox_output_path()}.}

\item{repo_name}{character. Name of the repository. Defaults to the name of the
current git repository found via \code{get_git_repo_name()}.}

\item{repos_subdir}{character. Subdirectory within the Dropbox directory where
the repository is stored. Defaults to "projects".}

\item{dropbox_base_dir}{character. Path to the base Dropbox directory. Defaults to
the Dropbox directory found via \code{find_dropbox_dir()}.}

\item{create_dir}{logical. Whether to create the directory if it does not exist.
Defaults to \code{FALSE}.}
}
\value{
A string with the path to the users Dropbox directory
}
\description{
This function tries to find the users Dropbox directory by looking for the
info.json file in the Dropbox directory of the users appdata.
}
