% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-calcs.R
\name{stats-calcs}
\alias{stats-calcs}
\alias{normalise}
\alias{scale_minmax}
\title{Statistical calculations}
\usage{
normalise(x)

scale_minmax(x, mn, mx)
}
\arguments{
\item{x}{numeric vector}

\item{mn}{numeric scalar. Minimum value of scaled output.}

\item{mx}{numeric scalar. Maximum value of scaled output.}
}
\description{
Various stastical calculations.
}
\section{Functions}{
\itemize{
\item \code{normalise}: normalise vector to max = 1 and min = 0

\item \code{scale_minmax}: normalise vector so that max = \code{mx} and min = \code{mn}
}}

