% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-calcs.R
\name{stats-calcs}
\alias{stats-calcs}
\alias{normalise}
\alias{scale_minmax}
\alias{max_z_given_n}
\alias{se}
\alias{lm_r2}
\alias{mad}
\alias{log10}
\alias{sym_setdiff}
\alias{q1}
\alias{q3}
\alias{p_se}
\title{Statistical calculations}
\usage{
normalise(x, in_min = min(x, na.rm = TRUE), in_max = max(x, na.rm = TRUE))

scale_minmax(
  x,
  out_min,
  out_max,
  in_min = min(x, na.rm = TRUE),
  in_max = max(x, na.rm = TRUE)
)

max_z_given_n(n)

se(x, ...)

lm_r2(lm_mod, adjusted = FALSE)

mad(x, na.rm = TRUE)

log10(x)

sym_setdiff(a, b)

q1(x)

q3(pmf)

p_se(p_est, k)
}
\arguments{
\item{x}{numeric vector}

\item{in_min}{minimum possible value of input, defaults to min(x, na.rm = TRUE)}

\item{in_max}{maximum possible value of input, defaults to max(x, na.rm = TRUE)}

\item{out_min}{numeric scalar. Minimum value of scaled output.}

\item{out_max}{numeric scalar. Maximum value of scaled output.}

\item{n}{numeric scalar. sample size.}

\item{lm_mod}{model output from stats::lm}

\item{adjusted}{logical. Whether to return adjusted R (\code{TRUE}), or non-adjusted
(\code{FALSE}). Defaults \code{FALSE}}

\item{p_est, k}{estimated p value from Fisherian randomization inference from K iterations}
}
\description{
Various stastical calculations.
}
\section{Functions}{
\itemize{
\item \code{normalise()}: normalise vector to scale [0, 1]

\item \code{scale_minmax()}: normalise vector so that max = \code{mx} and min = \code{mn}

\item \code{max_z_given_n()}: Find Maximum z score given sample size n

\item \code{se()}: Calculate standard error of mean

\item \code{lm_r2()}: get r squared from stats::lm()

\item \code{mad()}: get median absolute deviation

\item \code{log10()}: Log(x, base = 10) wrapper

\item \code{sym_setdiff()}: symmetrical setdiff()

\item \code{q1()}: quantile wrapper for Q1

\item \code{q3()}: quantile wrapper for Q3

\item \code{p_se()}: estimate standard error of estimate p value from Fisherian randomization inference

}}
\seealso{
https://www.tandfonline.com/doi/abs/10.1080/00031305.1988.10475530
}
