% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{any0}
\alias{any0}
\alias{all0}
\alias{\%||\%}
\alias{is_wholenum}
\alias{brackets}
\title{Utility functions}
\usage{
any0(...)

all0(...)

x \%||\% y

is_wholenum(x)

brackets(x, type = c("smooth", "square", "squiggly"))
}
\arguments{
\item{...}{logical expression passed to \code{any()} or \code{all()}}
}
\description{
Set of simple utility functions for common use.
}
\section{Functions}{
\itemize{
\item \code{any0}: wrapper for \code{any()} with na.rm = TRUE

\item \code{all0}: wrapper for \code{all()} with na.rm = TRUE

\item \code{\%||\%}: replace \code{NULL} with default value

\item \code{is_wholenum}: check if values are whole numbers

\item \code{brackets}: enclose string in brackets.
}}

